% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrondissements_geojson.R
\docType{data}
\name{arrondissements_geojson}
\alias{arrondissements_geojson}
\title{arrondissements_geojson data base}
\description{
arrondissements_geojson data base
}
\author{
Alex \email{alexandre.momeni@polytechnique.edu}
}
\keyword{data}
