% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Map.R
\docType{data}
\name{Data_Map}
\alias{Data_Map}
\title{Data_Map : duplicate}
\description{
Data_Map : duplicate
}
\references{
\url{https://opendata.paris.fr/explore/dataset/utilisations_mensuelles_des_hotspots_paris_wi-fi/?q=&sort=stop_time}{The original link}
}
\author{
Alex \email{alexandre.momeni@polytechnique.edu}
}
\keyword{data}
