% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Data_Map.R
\name{Data_Map_Filter}
\alias{Data_Map_Filter}
\title{Filter_Data_Map}
\usage{
Data_Map_Filter(start, end, duration_min = 0, duration_max = 7200,
  districts = c("All"), cat_sites = c("All"), sites = c("All"),
  countries = c("All"), devices = c("All"))
}
\arguments{
\item{start}{= a POSIXct, end = a POSIXct, duration_min = an int (default is 0), duration_max = an int (default is 7200), districts, = a vector (default is c("All")) , cat_sites = idem, sites = idem, countries = idem, devices = idem}
}
\value{
a dataframe
}
\description{
Filter the wifi_connexion_data DB joined with the mapping DB, according to the filters parametrized in the Shiny App
}
