% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Viz_1.R
\docType{data}
\name{Data_Viz_1}
\alias{Data_Viz_1}
\title{Data_Viz_1 : The wifi_connexion_data DB processed to fit its utilisation for the 1st visualization}
\description{
Data_Viz_1 : The wifi_connexion_data DB processed to fit its utilisation for the 1st visualization
}
\author{
Sacha \email{sacha.izadi@polytechnique.edu}
}
\keyword{data}
