% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_country.R
\docType{data}
\name{mapping_country}
\alias{mapping_country}
\title{mapping_country region data base : each country is linked to a region of the world}
\description{
mapping_country region data base : each country is linked to a region of the world
}
\author{
Amir \email{amir.benmahjoub@polytechnique.edu}
}
\keyword{data}
