% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_devices.R, R/mapping_langues.R,
%   R/mapping_sites_gps_catsite.R
\docType{data}
\name{mapping_devices}
\alias{mapping_devices}
\alias{mapping_devices}
\alias{mapping_devices}
\title{mapping_devices data base : each group (browser, os, device) defines a category of device (smartphone, tablet, computer, other)}
\description{
mapping_devices data base : each group (browser, os, device) defines a category of device (smartphone, tablet, computer, other)

mapping_langues data base

mapping_sites_gps_catsite data base
}
\author{
Sacha \email{sacha.izadi@polytechnique.edu}

Amir \email{amir.benmahjoub@polytechnique.edu}

Amir \email{amir.benmahjoub@polytechnique.edu}
}
\keyword{data}
