% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nb_connexions.R
\name{number_connexions}
\alias{number_connexions}
\title{Returns the number of connexions of a set vs. time with respects to a given axis of analysis}
\usage{
number_connexions(set, start, end, analysis_axis)
}
\arguments{
\item{a_set}{= a dataframe, start = a POSXct, end = a POSXct, analysis_axis = a string of char to chose in ("None","Country","category_device","site","Ardt","category_site")}
}
\value{
a dataframe
}
\description{
Returns the number of connexions of a set vs. time with respects to a given axis of analysis
}
\examples{
\dontrun{
start<-ymd("2016-09-10")
end<-ymd("2016-09-17")
duration_max<- 5000
districts<-c(1,2,5,8,"All")
districts<-c(1,2,5,8)
cat_sites<-c("Mairie","Bibliothèque")
sites<-c("All")
countries<-c("All")
devices<-c("smartphone","tablet")
analysis_axis<-"category_device"
a_set <- Viz1_Filter(start, end, duration_max, districts, cat_sites, sites, countries, devices)

number_connexions(a_set,start,end, analysis_axis)
}


}
