% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz1_plot.R
\name{Viz1_plot}
\alias{Viz1_plot}
\title{Plots the number of wifi connexions with respects to a given parametrization}
\usage{
Viz1_plot(start, end, duration_min = 0, duration_max = 7200,
  districts = c("All"), cat_sites = c("All"), sites = c("All"),
  countries = c("All"), devices = c("All"), analysis_axis = "None",
  complete_data = FALSE)
}
\arguments{
\item{start}{= a POSXct, end = a POSXct, duration_min = an int (default is 0), duration_max = an int (default=7200), districts = a vector (default=c("All")), cat_sites = idem, sites = idem, countries = idem, devices = idem, analysis_axis = a string of char to chose in ("None","Country","category_device","site","Ardt","category_site" - default="None"), complete_data = Boolean (default=FALSE)}
}
\value{
a dataframe
}
\description{
Plots the number of wifi connexions with respects to a given parametrization
}
\examples{
\dontrun{
start<-lubridate::ymd("2016-09-27")
end<-lubridate::ymd("2016-09-30")
duration_min<- 10
duration_max<- 5000
districts<-c(1,2,5,8,"All")
cat_sites<-c("Mairie","Bibliothèque")
sites<-c("All")
countries<-c("All")
devices<-c("smartphone","tablet")
analysis_axis<-"category_device"

Viz1_plot(start, end, duration_min, duration_max, districts, cat_sites, sites, countries, devices, analysis_axis, TRUE)
}


}
