% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz1_plot_V2.R
\name{Viz1_plot_V2}
\alias{Viz1_plot_V2}
\title{Plots the number of wifi connexions with respects to a given parametrization}
\usage{
Viz1_plot_V2(a_set, start, end, analysis_axis = "None",
  complete_data = FALSE)
}
\arguments{
\item{a_set}{= a dataset, start = a POSXct, end = a POSXct, duration_min = an int (default is 0), duration_max = an int (default=7200), districts = a vector (default=c("All")), cat_sites = idem, sites = idem, countries = idem, devices = idem, analysis_axis = a string of char to chose in ("None","Country","category_device","site","Ardt","category_site" - default="None"), complete_data = Boolean (default=FALSE)}
}
\value{
a dataframe
}
\description{
Plots the number of wifi connexions with respects to a given parametrization
}
\examples{
\dontrun{
start<-lubridate::ymd("2016-09-27")
end<-lubridate::ymd("2016-09-30")
duration_min<- 10
duration_max<- 5000
districts<-c(1,2,5,8,"All")
cat_sites<-c("Mairie","Bibliothèque")
sites<-c("All")
countries<-c("All")
devices<-c("smartphone","tablet")
analysis_axis<-"category_device"
a_set <- Viz1_Filter(start, end, duration_min, duration_max, districts, cat_sites, sites, countries, devices) \%>\%
number_connexions(start,end,analysis_axis)

Viz1_plot_V2(a_set, start, end, analysis_axis, TRUE)
}


}
