% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz3_plot.R
\name{Viz3_plot}
\alias{Viz3_plot}
\title{Plots a treemap with size = to average number of connexions & color = connexion duration with respect to a given analysis axis}
\usage{
Viz3_plot(a_set, analysis_axis = "None", hierarchy_view)
}
\arguments{
\item{a_set}{= a dataframe, analysis_axis = a defined string of characters (default is "None"), hierarchy_view = a boolean}
}
\value{
a treemap object
}
\description{
Plots a treemap with size = to average number of connexions & color = connexion duration with respect to a given analysis axis
}
\examples{
\dontrun{

A MODIFIER

start<-lubridate::ymd("2016-09-27")
end<-lubridate::ymd("2016-09-30")
duration_min<- 10
duration_max<- 5000
districts<-c(1,2,5,8,"All")
cat_sites<-c("Mairie","Bibliothèque")
sites<-c("All")
countries<-c("All")
devices<-c("smartphone","tablet")
analysis_axis<-"category_device"
hierarchy_view<-FALSE
a_set <- Viz1_Filter(start, end, duration_min, duration_max, districts, cat_sites, sites, countries, devices) \%>\%
number_connexions(start,end,analysis_axis)

Viz3_plot(a_set, analysis_axis, hierarchy_view)

}

}
