% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wifi_connexion_data.R
\docType{data}
\name{wifi_connexion_data}
\alias{wifi_connexion_data}
\title{wifi_connexion_data : The cleaned version of the database donwloaded from the Paris Open Data website}
\description{
wifi_connexion_data : The cleaned version of the database donwloaded from the Paris Open Data website
}
\references{
\url{https://opendata.paris.fr/explore/dataset/utilisations_mensuelles_des_hotspots_paris_wi-fi/?q=&sort=stop_time}{The original link}
}
\author{
Sacha \email{sacha.izadi@polytechnique.edu}
}
\keyword{data}
