% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help.R
\docType{package}
\name{wifiparis}
\alias{wifiparis}
\alias{wifiparis-package}
\title{Wifi Package}
\description{
The package has the functions and datasets underlying the Shiny App developped for the wifiparis project
}
\author{
Amir Benmahjoub <amir.benmahjoub@polytechnique.edu>, Alexandre Momeni <alexandre.momeni@polytechnique.edu>, Sacha Izadi <sacha.izadi@polytechnique.edu>
}
