% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_plot.R
\name{Map_plot}
\alias{Map_plot}
\title{Map_Plot}
\usage{
Map_plot(start, end, duration_min = 0, duration_max = 7200,
  districts = c("All"), cat_sites = c("All"), sites = c("All"),
  countries = c("All"), devices = c("All"))
}
\arguments{
\item{start}{= a POSXct, end = a POSXct, duration_min = an int (default is 0), duration_max = an int (default=7200), districts = a vector (default=c("All")), cat_sites = idem, sites = idem, countries = idem, devices = idem}
}
\value{
a dataframe
}
\description{
Plots filtered data onto an Open Street Map of Paris
}
