% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Plot_Backup.R, R/Map_plot.R
\name{Map_plot}
\alias{Map_plot}
\alias{Map_plot}
\title{Plots a map}
\usage{
Map_plot(a_set)

Map_plot(a_set)
}
\arguments{
\item{a_set}{= a dataframe}

\item{start}{= a POSXct, end = a POSXct, duration_min = an int (default is 0), duration_max = an int (default=7200), districts = a vector (default=c("All")), cat_sites = idem, sites = idem, countries = idem, devices = idem}
}
\value{
a dataframe

a leaflet
}
\description{
Plots a map

Plots a map
}
\examples{
\dontrun{
A FAIRE
}

\dontrun{
start<-lubridate::ymd("2016-09-27")
end<-lubridate::ymd("2016-09-30")
duration_min<- 0
duration_max<- 7200
districts<-c("20","11","18")
cat_sites<-c("Bibliothèque")
sites<-c("All")
countries<-c("All")
devices<-c("smartphone","table")
a_set<-Viz1_Filter(start, end, duration_min, duration_max, districts, cat_sites, sites, countries, devices)

Map_plot(a_set)

}

}
