% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Data_Viz_1.R
\name{Viz1_Filter}
\alias{Viz1_Filter}
\title{Filter the wifi_connexion_data DB joined with the mapping DB, according to the filters parametrized in the Shiny App}
\usage{
Viz1_Filter(start, end, duration_max = 7200, districts = c("All"),
  cat_sites = c("All"), sites = c("All"), countries = c("All"),
  devices = c("All"))
}
\arguments{
\item{start}{= a POSIXct, end = a POSIXct, duration_max = an int (default is 7200), districts, = a vector (default is c("All")) , cat_sites = idem, sites = idem, countries = idem, devices = idem}
}
\value{
a dataframe
}
\description{
Filter the wifi_connexion_data DB joined with the mapping DB, according to the filters parametrized in the Shiny App
}
\examples{
\dontrun{
start<-ymd("2016-09-10")
end<-ymd("2016-09-12")
duration_max<- 5000
districts<-c(1,2,5,8,"All")
cat_sites<-c("Mairie","Bibliothèque")
sites<-c("Discothèque des Halles","Bibliothèque Port Royal","Bibliothèque Mouffetard")
countries<-c("SPAIN","FRANCE")
devices<-c("smartphone","tablet")
Viz1_Filter(start, end, duration_max, districts, cat_sites, sites, countries, devices) \%>\% View()
}

}
